big changes. made all the KNN's predict one point. that should make our life easier for parameter tuning. also made it so that we make PointSummaries whenever we are classifying points. Testing accuracy is simple now, but seems like there may be an issue somewhere. accuracy is down and I can't tell why. Might be the KNN stuff, because right now every single KNN matrix is showing 0 of each class for some reason.

+++ data_utilities/StatStructs.h
@@ -13,7 +13,7 @@
 struct BlockInfo {
     int blockClass;
     int blockIdx;
-    float blockSize;
+    int blockSize;
     float blockDensity;
 };
